Move Rig and Animation loading to gltf/LoadAnimation.
