+++ src/apps/skinning_sandbox/skinning_sandbox/Scene.cpp
@@ -18,7 +18,7 @@
 #include <snac-renderer/ResourceLoad.h>
 
 #include <snac-renderer/gltf/GltfLoad.h>
-#include <snac-renderer/gltf/LoadBuffer.h>
+#include <snac-renderer/gltf/LoadAnimation.h>
 
 
 namespace ad {
@@ -27,227 +27,12 @@ namespace snac {
 
 const std::string gDebugDrawer = "skinning";
 
-// Internal data structure used while loading the gltf animation
-struct Anim
-{
-    using Index_t = arte::gltf::Index<arte::gltf::Accessor>;
-    static constexpr Index_t gInvalidIndex = std::numeric_limits<Index_t::Value_t>::max();
-    using Path = arte::gltf::animation::Target::Path;
-
-    Anim(arte::gltf::Index<arte::gltf::Node> aTargetNode) :
-        mGltfNode{aTargetNode}
-    {}
-
-    /// @return The index of the input (time) accessor.
-    arte::gltf::Index<arte::gltf::Accessor>
-    add(Path aPath, arte::Const_Owned<arte::gltf::animation::Sampler> aSampler)
-    {
-        // Only support linear interpolation at the moment.
-        assert(aSampler->interpolation == arte::gltf::animation::Sampler::Interpolation::Linear);
-
-        auto write = [](Index_t & aTarget, Index_t aValue)
-        {
-            if(aTarget != gInvalidIndex)
-            {
-                throw std::logic_error{"Duplicated node path in an animation."};
-            }
-            else
-            {
-                aTarget = aValue;
-            }
-        };
-
-        switch(aPath)
-        {
-            case Path::Translation:
-                write(translationAccessor, aSampler->output);
-                break;
-            case Path::Rotation:
-                write(rotationAccessor, aSampler->output);
-                break;
-            case Path::Scale:
-                write(scaleAccessor, aSampler->output);
-                break;
-            default:
-                throw std::logic_error{"Unsupported animation target path."};
-        }
-
-        return aSampler->input;
-    }
-
-    bool isComplete() const
-    {
-        return translationAccessor != gInvalidIndex
-            && rotationAccessor != gInvalidIndex
-            && scaleAccessor != gInvalidIndex;
-    }
-
-    // Only used if there is no sampler for a path:
-    // The value is then copied from the gltf node
-    arte::gltf::Index<arte::gltf::Node> mGltfNode;
-    Index_t translationAccessor = gInvalidIndex;
-    Index_t rotationAccessor = gInvalidIndex;
-    Index_t scaleAccessor = gInvalidIndex;
-};
-
-
-void assertAccessor(arte::Const_Owned<arte::gltf::Accessor> aAccessor, 
-                    arte::gltf::Accessor::ElementType aElementType,
-                    arte::gltf::EnumType aComponentType) 
-{
-    assert(aAccessor->type == aElementType);
-    assert(aAccessor->componentType == aComponentType);
-}
-
-
-NodeAnimation readAnimation(arte::Const_Owned<arte::gltf::Animation> aGltfAnimation,
-                            const std::vector<Node::Index> & aGltfToTreeIndex,
-                            const arte::Gltf & aGltf)
-{
-    std::unordered_map<std::size_t/*nodetree index*/, Anim> perNode;
-
-    Anim::Index_t inputAccessorIdx = Anim::gInvalidIndex;
-
-    auto samplers = aGltfAnimation.iterate(&arte::gltf::Animation::samplers);
-    for (const arte::gltf::animation::Channel & channel : aGltfAnimation->channels)
-    {
-        assert(channel.target.node.has_value());
-        
-        // Get the existing Anim for this nodetree index, or insert an Anim if its is not present.
-        auto [iterator, didInsert] = perNode.try_emplace(
-            aGltfToTreeIndex[*channel.target.node], // Key
-            *channel.target.node // gltf node index
-        );
-        auto samplerInput = iterator->second.add(channel.target.path, samplers.at(channel.sampler));
-
-        if(inputAccessorIdx != Anim::gInvalidIndex && inputAccessorIdx != samplerInput)
-        {
-            throw std::logic_error{"Expecting all channels of a given animation to have the same input."};
-        }
-        else
-        {
-            inputAccessorIdx = samplerInput;
-        }
-    }
-
-    NodeAnimation animation;
-
-    auto inputAccessor = aGltfAnimation.get(inputAccessorIdx);
-    assertAccessor(inputAccessor, arte::gltf::Accessor::ElementType::Scalar, GL_FLOAT);
-    animation.mTimepoints = loadTypedData<float>(inputAccessor);
-    animation.mEndTime = std::get<arte::gltf::Accessor::MinMax<float>>(*inputAccessor->bounds).max[0];
-    for(const auto & [nodeIndex, anim] : perNode)
-    {
-        // For each path (rotation, translation, scale), read the keyframe data if the path is animated
-        // or make N copies of the node's initial value for the path if it is not animated.
-        // TODO #anim Can we not make N copies of the constant channels?
-
-        NodeAnimation::NodeKeyframes nodeKeyframes;
-        auto gltfNode = aGltf.get(anim.mGltfNode);
-
-        if (anim.translationAccessor != Node::gInvalidIndex)
-        {
-            auto translationAccessor = aGltfAnimation.get(anim.translationAccessor);
-            assertAccessor(translationAccessor, arte::gltf::Accessor::ElementType::Vec3, GL_FLOAT);
-            nodeKeyframes.mTranslations = loadTypedData<math::Vec<3, float>>(translationAccessor);
-        }
-        else
-        {
-            std::fill_n(std::back_inserter(nodeKeyframes.mTranslations),
-                        animation.mTimepoints.size(),
-                        getTransformationAsTRS(gltfNode).translation);
-        }
-
-        if (anim.rotationAccessor != Node::gInvalidIndex)
-        {
-            auto rotationAccessor = aGltfAnimation.get(anim.rotationAccessor);
-            assertAccessor(rotationAccessor, arte::gltf::Accessor::ElementType::Vec4, GL_FLOAT);
-            nodeKeyframes.mRotations = loadTypedData<math::Quaternion<float>>(rotationAccessor);
-        }
-        else
-        {
-            std::fill_n(std::back_inserter(nodeKeyframes.mRotations),
-                        animation.mTimepoints.size(),
-                        getTransformationAsTRS(gltfNode).rotation);
-        }
-
-        if (anim.scaleAccessor != Node::gInvalidIndex)
-        {
-            auto scaleAccessor = aGltfAnimation.get(anim.scaleAccessor);
-            assertAccessor(scaleAccessor, arte::gltf::Accessor::ElementType::Vec3, GL_FLOAT);
-            nodeKeyframes.mScales = loadTypedData<math::Vec<3, float>>(scaleAccessor);
-        }
-        else
-        {
-            std::fill_n(std::back_inserter(nodeKeyframes.mScales),
-                        animation.mTimepoints.size(),
-                        getTransformationAsTRS(gltfNode).scale);
-        }
-
-        animation.mNodes.push_back(nodeIndex);
-        animation.mKeyframesEachNode.push_back(std::move(nodeKeyframes));
-    }
-
-    return animation;
-}
-
-
-std::unordered_map<std::string, NodeAnimation>
-loadAnimations(const arte::Gltf & aGltf,
-               const std::vector<Node::Index> & aGltfToTreeIndex)
-{
-    std::unordered_map<std::string, NodeAnimation> animations;
-
-    for(arte::Const_Owned<arte::gltf::Animation> animation : aGltf.getAnimations())
-    {
-        std::string name = animation->name;
-        name = (name.empty() ? "anim_#" + animation.id() : name);
-        animations.emplace(name, readAnimation(animation, aGltfToTreeIndex, aGltf));
-    }
-
-    return animations;
-}
-
-
-Rig loadSkin(const arte::Gltf & aGltf,
-             const std::vector<Node::Index> & aGltfToTreeIndex,
-             arte::gltf::Index<arte::gltf::Skin> aSkinIndex = 0)
-{
-    arte::Const_Owned<arte::gltf::Skin> skin = aGltf.get(aSkinIndex);
-
-    Rig result{
-        .mName = skin->name,
-    };
-
-    for (auto gltfNodeIdx : skin->joints)
-    {
-        result.mJoints.push_back(aGltfToTreeIndex[gltfNodeIdx]);
-    }
-
-    // Only support skin with inverse bind matrices
-    assert(skin->inverseBindMatrices);
-    auto ibmAccessor = skin.get(*skin->inverseBindMatrices);
-    assertAccessor(ibmAccessor, arte::gltf::Accessor::ElementType::Mat4, GL_FLOAT);
-    result.mInverseBindMatrices = loadTypedData<math::AffineMatrix<4, float>>(ibmAccessor);
-
-    return result;
-}
-
 
 Scene::RiggingData loadRiggingData(const arte::Gltf & aGltf)
 {
     Scene::RiggingData result;
-    std::vector<Node::Index> indexMapping; // gltf to tree
-    NodeTree<arte::gltf::Node::Matrix> scene;
-
-    // TODO #anim should only load the joints instead of the whole scene.
-    std::tie(scene, indexMapping) = loadHierarchy(aGltf, 0/*scene idx*/);
-
-    result.mRig = loadSkin(aGltf, indexMapping, 0);
-    result.mRig.mScene = std::move(scene);
-
-    result.mAnimations = loadAnimations(aGltf, indexMapping);
-
+    std::tie(result.mRig, result.mAnimations) =
+        loadSkeletalAnimation(aGltf, 0/*skin index*/, 0/*scene index*/);
     return result;
 }
