+++ src/libs/snac-renderer/snac-renderer/Mesh.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "IntrospectProgram.h"
+#include "Rigging.h"
 #include "Semantic.h"
 #include "UniformParameters.h"
