+++ src/libs/snac-renderer/snac-renderer/gltf/GltfLoad.h
@@ -29,13 +29,6 @@ inline Model loadGltf(filesystem::path aModel)
     return loadGltf(aModel, aModel.filename().string());
 }
 
-/// @return A pair with first being the tree, where parents are guaranteed to happen 
-/// **before** their children.
-/// Second is an array mapping the node index in the gltf (array index) 
-/// to an index in the node tree (array value).
-std::pair<NodeTree<arte::gltf::Node::Matrix>, std::vector<Node::Index>>
-loadHierarchy(const arte::Gltf & aGltf, std::size_t aSceneIndex = 0);
-
 
 } // namespace snac
 } // namespace ad
